+++ AppCore/.gitignore
