+++ AppCore/Package.swift
@@ -1,0 +1,33 @@
+// swift-tools-version:5.5
+// The swift-tools-version declares the minimum version of Swift required to build this package.
+
+import PackageDescription
+
+let package = Package(
+    name: "AppCore",
+    platforms: [
+        .iOS(.v15)
+    ],
+    products: [
+        // Products define the executables and libraries a package produces, and make them visible to other packages.
+        .library(
+            name: "AppCore",
+            targets: ["AppCore"]),
+    ],
+    dependencies: [
+        .package(name: "Kanna", url: "https://github.com/tid-kijyun/Kanna.git", from: "5.2.2")
+    ],
+    targets: [
+        // Targets are the basic building blocks of a package. A target can define a module or a test suite.
+        // Targets can depend on other targets in this package, and on products in packages this package depends on.
+        .target(
+            name: "AppCore",
+            dependencies: [
+                "Kanna"
+            ]
+        ),
+        .testTarget(
+            name: "AppCoreTests",
+            dependencies: ["AppCore"]),
+    ]
+)
