+++ AppCore/Sources/AppCore/API/YaeyamaKankouApi.swift
