+++ AppCore/Sources/AppCore/Model/RouteScheduleListResponse.swift
