+++ AppCore/Sources/AppCore/Model/RouteStatusListResponse.swift
